update booking flow logic
