+++ package.json
@@ -10,6 +10,7 @@
     "antd": "^4.20.4",
     "axios": "^0.27.2",
     "js-cookie": "^3.0.1",
+    "moment": "^2.29.3",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-redux": "^8.0.1",
