+++ src/constants/RoleConstants.js
@@ -2,6 +2,6 @@ export const RoleConstant = {
   MANAGER: "MANAGER",
   CUSTOMER_SERVICE: "CUSTOMER_SERVICE",
   DENTIST: "DENTIST",
-  CUSTOMER: "CUSTOMER",
+  PATIENT: "PATIENT",
   ADMIN: "ADMIN",
 };
