+++ src/utils/convert.utils.js
@@ -1,0 +1,7 @@
+export const convertMomentToDate = (momentData) => {
+  return new Date().setMilliseconds(momentData.utc().valueOf());
+};
+
+export const convertMomentToMilliseconds = (momentData) => {
+  return momentData.utc().valueOf();
+};
